requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
lxml
MarkupPy==1.14

multitasking==0.0.9
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7

Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn
scipy==1.6.3
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
yfinance==0.1.63
django-common-helpers==0.9.2
django-cron==0.5.1
